Initial commit for the `camfwd` project
This is the first commit for the `camfwd` project. The objective of this
project is to produce a tool for interacting with DSLRs (and other
cameras supported by `libgphoto2`) from the command line. This tool will
be able to forward live output from the camera to another program, such
as a video player, and will offer an interface for quickly manipulating
camera parameters and focus.

As it stands in this commit, the tool is currently capable of forwarding
output from the default camera to a program of the user's choosing. It
can be started by executing the `camfwd.camfwd` module and specifying
the program/command to which output should be forwarded in the arguments
to the tool.