Introduce a module for command-line interface construction
The new KeyAct class from the `act.py` module is intended to help with
constructing a command-line interface for the rest of the project. When
started, it sets the correct terminal modes to disable echo and enable
single key inputs. Then, the `process_input()` method can be repeatedly
called to interpret any keys that are pressed and execute the actions
specified by the `key_map` dictionary from the constructor. There are
some caveats to this class' behavior, so a more elegant solution may be
devised in the future.