Add the ability to control exposure settings via key commands
Users will now be able to control the shutter speed, focal ratio
(f-number), and film speed (ISO) via a series of key commands. This new
functionality makes use of the `act` module included in the project, and
a call to the `process_input()` method of a KeyAct object has been
placed within the main control loop. When a key is pressed, the setting
will be changed, and the new value will be printed to the screen. The
key controls are laid out logically on the right side of the keyboard,
but they are not currently configurable. This may be changed in a future
version.