Add controls to permit the camera's focus to be adjusted
A camera connected to `camfwd` can now have its focus adjusted in steps
of three different sizes via key commands. The keys are laid out on the
left side of keyboard in pairs. One key in each pair is set to move the
focus toward infinity, and the other key moves the focus away from
infinity. Due differences in command sets and the large variety of
camera models available, it is difficult to determine whether the design
of this code will work correctly in all cases, but it remains flexible
in case adjustments need to be made.