Introduce unit tests for the project
Using Python's built-in `unittest` module, this commit introduces unit
tests for the project. The tests can be automatically discovered and run
by executing `python3 -m unittest` in the root directory of the project.
Most of the code in the project is tested, with the exception of a few
aspects that would be impossible or extremely difficult to test in an
automated way.

The addition of these tests necessitated several changes to the
program's source code. Namely, the I/O operations of the KeyAct class
had to be factored out into the new StandardIO class so that the KeyAct
class could be tested using the new InternalIO class. Given that Python
has no facilities for emulating a TTY, the StandardIO class cannot be
tested. However, the InternalIO class is designed to only utilize
built-in Python facilities, thus allowing the KeyAct class to be tested.

The CameraControl class was also updated to improve its testability. Its
initialization and tear-down methods were exposed as normal methods so
that it can be used without a `with` statement. The initialization
procedure was also modified so that the camera is always placed in
"viewfinder" mode, thus allowing the focus to be adjusted even if no
output is captured from the camera.