Add the ability to measure contrast and entropy
The `camfwd` tool can now measure the current contrast and entropy of a
camera's output. When the proper key is pressed, a frame of live view
output will be captured and processed using the `Pillow` library to
calculate either the contrast or the entropy. The contrast measurement
should help a user deduce whether their image is in focus, and the
entropy measurement should help a user deduce the amount of noise in
their image.