Introduce a key control to trigger image capture
This commit introduces a new key control to prompt a connected camera to
capture an image and save said image to its internal storage. When the
'Enter' key is pressed, `camfwd` sets the destination for the image
capture to the camera's storage, triggers an image capture, and prints a
brief description of the location to which the capture was saved.