Add unit tests for image analysis and camera image capture
This commit introduces unit tests for the new image analysis and image
capture code. The image contrast and entropy functions are tested
against a bitmap image with known values for these two calculations. The
image conversion function from the `analysis` module is also tested
using a JPEG image embedded in the code as a Base64 string. The image
capture code is tested by prompting an image capture, performing a crude
test to determine whether it was saved to the camera's storage, reading
the image into memory, and checking whether it appears to be a JPEG.